changes mentioned in PR & code refactoring
